[FLINK-15053][runtime] Escape all dynamic property values for taskmanager
For unix-like OS(Linux, MacOS, FREE_BSD, etc.), each value will put in single quotes. This works for all chars except single quote itself. To escape the single quote, close the quoting before it, insert the escaped single quote, and then re-open the quoting. For example, the value is foo'bar and the escaped value is 'foo'\''bar'. Navigate to the following url for more information.

https://stackoverflow.com/questions/15783701/which-characters-need-to-be-escaped-when-using-bash