[FLINK-15836][k8s] Start a new pods watcher in KubernetesResourceManager when the old one is closed with exception
By default the watcher will always reconnect in Kubernetes client internally. However, if the watchReconnectLimit is configured by users via java properties or environment, the watcher may be stopped. Then all the changes of pods will not be processed properly. So we need to start a new pods watcher in KubernetesResourceManager when the old one is closed with exception.