[hotfix][coordination] Introduce new LeaderElection and LeaderRetrieval composition mechanism and wire Zookeeper implementation to new interfaces
After this refactor, we will have a DefaultLeaderElectionService. Composed with different LeaderElectionDriver, we could perform a leader election for the contender, and then persist the leader information to various storage. Also we will get a DefaultLeaderRetrievalService. Composed with different LeaderRetrievalDriver, we could we could retrieve the leader information from different storages.

This will make the contract more clear and easier to add new leader election/retrieval services.